+++ src/cmdClientCreateGrow.go
@@ -5,12 +5,14 @@ import "os"
 type clientCreateCmd struct {
 	Base  clientCreateBaseCmd  `command:"base" subcommands-optional:"true" description:"simple base image"`
 	Tools clientCreateToolsCmd `command:"tools" subcommands-optional:"true" description:"aerospike-tools"`
+	AMS   clientCreateAMSCmd   `command:"ams" subcommands-optional:"true" description:"prometheus and grafana for AMS; for exporter see: cluster add exporter"`
 	// NEW_CLIENTS_CREATE
 	Help helpCmd `command:"help" subcommands-optional:"true" description:"Print help"`
 }
 
 type clientAddCmd struct {
 	Tools clientAddToolsCmd `command:"tools" subcommands-optional:"true" description:"aerospike-tools"`
+	AMS   clientAddAMSCmd   `command:"ams" subcommands-optional:"true" description:"prometheus and grafana for AMS; for exporter see: cluster add exporter"`
 	// NEW_CLIENTS_ADD
 	Help helpCmd `command:"help" subcommands-optional:"true" description:"Print help"`
 }
@@ -41,6 +43,11 @@ func init() {
 	addBackendSwitch("client.create.tools", "docker", &a.opts.Client.Create.Tools.Docker)
 	addBackendSwitch("client.grow.tools", "aws", &a.opts.Client.Grow.Tools.Aws)
 	addBackendSwitch("client.grow.tools", "docker", &a.opts.Client.Grow.Tools.Docker)
+
+	addBackendSwitch("client.create.ams", "aws", &a.opts.Client.Create.AMS.Aws)
+	addBackendSwitch("client.create.ams", "docker", &a.opts.Client.Create.AMS.Docker)
+	addBackendSwitch("client.grow.ams", "aws", &a.opts.Client.Grow.AMS.Aws)
+	addBackendSwitch("client.grow.ams", "docker", &a.opts.Client.Grow.AMS.Docker)
 
 	// NEW_CLIENTS_BACKEND